requirements.txt
libsass
jsmin
Jinja2
polars
Wikipedia
Biopython
requests
beautifulsoup4
lxml
